Move the materialized view freshness logic to TrinoCatalog
