+++ plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/IcebergMetadata.java
@@ -368,8 +368,6 @@ import static io.trino.plugin.iceberg.IcebergUtil.commit;
 import static io.trino.plugin.iceberg.IcebergUtil.createColumnHandle;
 import static io.trino.plugin.iceberg.IcebergUtil.deserializePartitionValue;
 import static io.trino.plugin.iceberg.IcebergUtil.fileName;
-import static io.trino.plugin.iceberg.IcebergUtil.firstSnapshot;
-import static io.trino.plugin.iceberg.IcebergUtil.firstSnapshotAfter;
 import static io.trino.plugin.iceberg.IcebergUtil.getColumnHandle;
 import static io.trino.plugin.iceberg.IcebergUtil.getColumnMetadatas;
 import static io.trino.plugin.iceberg.IcebergUtil.getCompressionPropertyName;
@@ -420,10 +418,6 @@ import static io.trino.spi.StandardErrorCode.QUERY_REJECTED;
 import static io.trino.spi.StandardErrorCode.TABLE_ALREADY_EXISTS;
 import static io.trino.spi.StandardErrorCode.TABLE_NOT_FOUND;
 import static io.trino.spi.StandardErrorCode.TYPE_MISMATCH;
-import static io.trino.spi.connector.MaterializedViewFreshness.Freshness.FRESH;
-import static io.trino.spi.connector.MaterializedViewFreshness.Freshness.FRESH_WITHIN_GRACE_PERIOD;
-import static io.trino.spi.connector.MaterializedViewFreshness.Freshness.STALE;
-import static io.trino.spi.connector.MaterializedViewFreshness.Freshness.UNKNOWN;
 import static io.trino.spi.connector.RetryMode.NO_RETRIES;
 import static io.trino.spi.connector.RowChangeParadigm.DELETE_ROW_AND_INSERT_ROW;
 import static io.trino.spi.predicate.TupleDomain.withColumnDomains;
@@ -4445,190 +4439,7 @@ public class IcebergMetadata
     @Override
     public MaterializedViewFreshness getMaterializedViewFreshness(ConnectorSession session, SchemaTableName materializedViewName, boolean considerGracePeriod)
     {
-        Optional<ConnectorMaterializedViewDefinition> materializedViewDefinition = getMaterializedView(session, materializedViewName);
-        if (materializedViewDefinition.isEmpty()) {
-            // View not found, might have been concurrently deleted
-            return new MaterializedViewFreshness(STALE, Optional.empty());
-        }
-
-        SchemaTableName storageTableName = materializedViewDefinition.get().getStorageTable()
-                .map(CatalogSchemaTableName::getSchemaTableName)
-                .orElseThrow(() -> new IllegalStateException("Storage table missing in definition of materialized view " + materializedViewName));
-
-        Table icebergTable = catalog.loadTable(session, storageTableName);
-        Optional<Snapshot> currentSnapshot = Optional.ofNullable(icebergTable.currentSnapshot());
-        String dependsOnTables = currentSnapshot
-                .map(snapshot -> snapshot.summary().getOrDefault(DEPENDS_ON_TABLES, ""))
-                .orElse("");
-        boolean dependsOnTableFunctions = currentSnapshot
-                .map(snapshot -> Boolean.valueOf(snapshot.summary().getOrDefault(DEPENDS_ON_TABLE_FUNCTIONS, "false")))
-                .orElse(false);
-        // For MVs refreshed before non-deterministic function tracking was added this flag
-        // defaults to false. Such MVs will be correctly flagged after their next refresh.
-        boolean dependsOnNonDeterministicFunctions = currentSnapshot
-                .map(snapshot -> Boolean.valueOf(snapshot.summary().getOrDefault(DEPENDS_ON_NON_DETERMINISTIC_FUNCTIONS, "false")))
-                .orElse(false);
-
-        Optional<Instant> refreshStartTime = currentSnapshot.map(snapshot -> snapshot.summary().get(TRINO_QUERY_START_TIME))
-                .map(Instant::parse);
-        Optional<Instant> refreshTime = refreshStartTime
-                // Fallback to snapshot commit time (end of refresh) for MVs defined before TRINO_QUERY_START_TIME was introduced
-                .or(() -> currentSnapshot.map(snapshot -> Instant.ofEpochMilli(snapshot.timestampMillis())));
-
-        if (dependsOnTableFunctions) {
-            // It can't be determined whether a value returned by table function is STALE or not
-            return new MaterializedViewFreshness(UNKNOWN, refreshTime);
-        }
-
-        if (dependsOnNonDeterministicFunctions) {
-            // Non-deterministic functions like current_timestamp produce different values over time,
-            // so the materialized view may be stale even if base tables haven't changed
-            return new MaterializedViewFreshness(UNKNOWN, refreshTime);
-        }
-
-        if (dependsOnTables.isEmpty()) {
-            // Information missing. While it's "unknown" whether storage is stale, we return "stale".
-            // Normally dependsOnTables may be missing only when there was no refresh yet.
-            return new MaterializedViewFreshness(STALE, Optional.empty());
-        }
-
-        Optional<java.time.Duration> gracePeriod = materializedViewDefinition.get().getGracePeriod();
-        if (considerGracePeriod && withinGracePeriod(session.getStart(), refreshStartTime, gracePeriod)) {
-            // To determine freshness, we normally load current metadata for each base table and check if there
-            // is a newer snapshot than the recorded one (DEPENDS_ON_TABLES). This requires expensive metastore
-            // operations for each base Iceberg table.
-            //
-            // The refresh query can read base table snapshots created before or during its execution. In the most
-            // pessimistic scenario, a new base table snapshot is created immediately after the refresh started
-            // (at refreshStartTime + epsilon), but the refresh reads an older snapshot. This new snapshot would
-            // not be recorded in DEPENDS_ON_TABLES, making the MV technically stale. However, when the caller set
-            // considerGracePeriod to true and refreshStartTime + gracePeriod > referenceTime, we can safely say that
-            // the MV is at least within the grace period because refreshStartTime is before the new snapshot creation time.
-            return new MaterializedViewFreshness(FRESH_WITHIN_GRACE_PERIOD, Optional.empty());
-        }
-
-        boolean hasUnknownTables = false;
-        OptionalLong firstTableChange = OptionalLong.of(Long.MAX_VALUE);
-        ImmutableList.Builder<Callable<TableChangeInfo>> tableChangeInfoTasks = ImmutableList.builder();
-        for (String tableToSnapShot : Splitter.on(',').split(dependsOnTables)) {
-            if (tableToSnapShot.equals(UNKNOWN_SNAPSHOT_TOKEN)) {
-                hasUnknownTables = true;
-                firstTableChange = OptionalLong.empty();
-                continue;
-            }
-
-            tableChangeInfoTasks.add(() -> getTableChangeInfo(session, tableToSnapShot));
-        }
-
-        boolean hasStaleIcebergTables = false;
-        List<TableChangeInfo> tableChangeInfos;
-
-        try {
-            tableChangeInfos = processWithAdditionalThreads(tableChangeInfoTasks.build(), metadataFetchingExecutor);
-        }
-        catch (ExecutionException e) {
-            throw new RuntimeException(e.getCause());
-        }
-
-        verifyNotNull(tableChangeInfos);
-
-        for (TableChangeInfo tableChangeInfo : tableChangeInfos) {
-            switch (tableChangeInfo) {
-                case NoTableChange() -> {
-                    // Fresh
-                }
-                case FirstChangeSnapshot(Snapshot snapshot) -> {
-                    hasStaleIcebergTables = true;
-                    firstTableChange = firstTableChange.isPresent() ?
-                            OptionalLong.of(Math.min(firstTableChange.getAsLong(), snapshot.timestampMillis())) :
-                            OptionalLong.empty();
-                }
-                case UnknownTableChange(), GoneOrCorruptedTableChange() -> {
-                    hasStaleIcebergTables = true;
-                    firstTableChange = OptionalLong.empty();
-                }
-            }
-        }
-
-        Optional<Instant> lastKnownFreshTime = firstTableChange.isPresent() ? Optional.of(Instant.ofEpochMilli(firstTableChange.getAsLong())) : refreshTime;
-        if (hasStaleIcebergTables) {
-            return new MaterializedViewFreshness(STALE, lastKnownFreshTime);
-        }
-        if (hasUnknownTables) {
-            return new MaterializedViewFreshness(UNKNOWN, lastKnownFreshTime);
-        }
-        return new MaterializedViewFreshness(FRESH, Optional.empty());
-    }
-
-    private boolean withinGracePeriod(Instant sessionStart, Optional<Instant> refreshStartTime, Optional<java.time.Duration> gracePeriod)
-    {
-        if (gracePeriod.isEmpty()) {
-            // infinite grace period
-            return true;
-        }
-        //noinspection OptionalIsPresent
-        if (refreshStartTime.isEmpty()) {
-            // refresh time unknown
-            return false;
-        }
-        return refreshStartTime.get().plus(gracePeriod.get()).isAfter(sessionStart);
-    }
-
-    private TableChangeInfo getTableChangeInfo(ConnectorSession session, String entry)
-    {
-        List<String> keyValue = Splitter.on("=").splitToList(entry);
-        if (keyValue.size() != 2) {
-            throw new TrinoException(ICEBERG_INVALID_METADATA, format("Invalid entry in '%s' property: %s'", DEPENDS_ON_TABLES, entry));
-        }
-        String tableName = keyValue.get(0);
-        String value = keyValue.get(1);
-        List<String> strings = Splitter.on(".").splitToList(tableName);
-        if (strings.size() == 3) {
-            strings = strings.subList(1, 3);
-        }
-        else if (strings.size() != 2) {
-            throw new TrinoException(ICEBERG_INVALID_METADATA, format("Invalid table name in '%s' property: %s'", DEPENDS_ON_TABLES, strings));
-        }
-        String schema = strings.get(0);
-        String name = strings.get(1);
-        SchemaTableName schemaTableName = new SchemaTableName(schema, name);
-        ConnectorTableHandle tableHandle = getTableHandle(session, schemaTableName, Optional.empty(), Optional.empty());
-
-        if (tableHandle == null || tableHandle instanceof CorruptedIcebergTableHandle) {
-            // Base table is gone or table is corrupted
-            return new GoneOrCorruptedTableChange();
-        }
-        OptionalLong snapshotAtRefresh;
-        if (value.isEmpty()) {
-            snapshotAtRefresh = OptionalLong.empty();
-        }
-        else {
-            snapshotAtRefresh = OptionalLong.of(Long.parseLong(value));
-        }
-        return getTableChangeInfo(session, (IcebergTableHandle) tableHandle, snapshotAtRefresh);
-    }
-
-    private TableChangeInfo getTableChangeInfo(ConnectorSession session, IcebergTableHandle table, OptionalLong snapshotAtRefresh)
-    {
-        Table icebergTable = catalog.loadTable(session, table.getSchemaTableName());
-        Snapshot currentSnapshot = icebergTable.currentSnapshot();
-
-        if (snapshotAtRefresh.isEmpty()) {
-            // Table had no snapshot at refresh time.
-            if (currentSnapshot == null) {
-                return new NoTableChange();
-            }
-            return firstSnapshot(icebergTable)
-                    .<TableChangeInfo>map(FirstChangeSnapshot::new)
-                    .orElse(new UnknownTableChange());
-        }
-
-        if (snapshotAtRefresh.getAsLong() == currentSnapshot.snapshotId()) {
-            return new NoTableChange();
-        }
-        return firstSnapshotAfter(icebergTable, snapshotAtRefresh.getAsLong())
-                .<TableChangeInfo>map(FirstChangeSnapshot::new)
-                .orElse(new UnknownTableChange());
+        return catalog.getMaterializedViewFreshness(session, materializedViewName, considerGracePeriod);
     }
 
     @Override
