+++ plugin/trino-iceberg/src/main/java/io/trino/plugin/iceberg/catalog/hms/TrinoHiveCatalogFactory.java
@@ -14,7 +14,6 @@
 package io.trino.plugin.iceberg.catalog.hms;
 
 import com.google.inject.Inject;
-import io.airlift.concurrent.BoundedExecutor;
 import io.trino.filesystem.TrinoFileSystemFactory;
 import io.trino.metastore.HiveMetastoreFactory;
 import io.trino.metastore.cache.CachingHiveMetastore;
@@ -33,9 +32,7 @@ import io.trino.spi.type.TypeManager;
 
 import java.util.Optional;
 import java.util.concurrent.Executor;
-import java.util.concurrent.ExecutorService;
 
-import static com.google.common.util.concurrent.MoreExecutors.directExecutor;
 import static io.trino.metastore.cache.CachingHiveMetastore.createPerTransactionCache;
 import static io.trino.plugin.iceberg.catalog.AbstractTrinoCatalog.TRINO_CREATED_BY_VALUE;
 import static java.util.Objects.requireNonNull;
@@ -67,7 +64,7 @@ public class TrinoHiveCatalogFactory
             IcebergTableOperationsProvider tableOperationsProvider,
             NodeVersion nodeVersion,
             @UsingSystemSecurity boolean isUsingSystemSecurity,
-            @ForIcebergMetadata ExecutorService metadataExecutorService)
+            @ForIcebergMetadata Executor metadataFetchingExecutor)
     {
         this.catalogName = requireNonNull(catalogName, "catalogName is null");
         this.metastoreFactory = requireNonNull(metastoreFactory, "metastoreFactory is null");
@@ -80,12 +77,7 @@ public class TrinoHiveCatalogFactory
         this.isUsingSystemSecurity = isUsingSystemSecurity;
         this.deleteSchemaLocationsFallback = config.isDeleteSchemaLocationsFallback();
         this.hideMaterializedViewStorageTable = config.isHideMaterializedViewStorageTable();
-        if (config.getMetadataParallelism() == 1) {
-            this.metadataFetchingExecutor = directExecutor();
-        }
-        else {
-            this.metadataFetchingExecutor = new BoundedExecutor(metadataExecutorService, config.getMetadataParallelism());
-        }
+        this.metadataFetchingExecutor = requireNonNull(metadataFetchingExecutor, "metadataFetchingExecutor is null");
     }
 
     @Override
